clean examples
